add retention duration SDK methods
I added SDK CRUD methods for interacting with soft-deleted
retention durations for a column. I also updated the userstore
sample to exercise a few of the methods.

Note that I did not add SDK methods for live retention durations,
since non-indefinite retention of live data is not yet supported,
and I did not add methods for tenant or purpose level defaults
of retention durations, since those are not yet exposed in console.